+++ web/menus/admin_menu.ex
@@ -8,8 +8,10 @@ defmodule Brando.Admin.Menu do
 
   menu %{
     name: gettext("Admin"), anchor: "admin", icon: "fa fa-dashboard icon",
-      submenu: [
-        %{name: gettext("Dashboard"), url: {:admin_dashboard_path, :dashboard}},
-        %{name: gettext("System info"), url: {:admin_dashboard_path, :system_info}, role: :superuser},
-        %{name: gettext("Log out"), url: {:session_path, :logout}}]}
+    submenu: [
+      %{name: gettext("Dashboard"), url: {:admin_dashboard_path, :dashboard}},
+      %{name: gettext("System info"), url: {:admin_dashboard_path, :system_info}, role: :superuser},
+      %{name: gettext("Log out"), url: {:session_path, :logout}}
+    ]
+  }
 end